refactor: viewer button, pagination button 로직 분리
