+++ scripts/pagination-button.js
@@ -1,0 +1,16 @@
+import { store } from "../store/index.js";
+import { nextPage, prevPage } from "../store/reducer.js";
+import { $nextPageButton, $prevPageButton } from "./doms.js";
+
+const handlePrevButtonClick = () => {
+  store.dispatch(prevPage());
+};
+
+const handleNextButtonClick = () => {
+  store.dispatch(nextPage());
+};
+
+export const addEventOnPaginationButton = () => {
+  $prevPageButton.addEventListener("click", handlePrevButtonClick);
+  $nextPageButton.addEventListener("click", handleNextButtonClick);
+};
